completely refactor the import cli

+++ db/migrations/20180630192209814_create_nutrition_fact.sql
@@ -1,10 +1,10 @@
 -- +micrate Up
 CREATE TABLE nutrition_facts (
   id BIGSERIAL PRIMARY KEY,
-  calories INT,
-  fat INT,
-  carbohydrate INT,
-  protein INT,
+  calories FLOAT,
+  fat FLOAT,
+  carbohydrate FLOAT,
+  protein FLOAT,
   created_at TIMESTAMP,
   updated_at TIMESTAMP
 );
